requirements.txt
aiohttp==3.8.1
aioredis==2.0.1
aiosignal==1.2.0
anyio==3.6.1
async-timeout==4.0.2
attrs==21.4.0
charset-normalizer==2.1.0
databases==0.6.0
fastapi==0.78.0
frozenlist==1.3.0
idna==3.3
loguru==0.6.0
multidict==6.0.2
Pillow==9.1.1
pydantic==1.9.1
PyYAML==6.0
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.19.1
typing-extensions==4.2.0
yarl==1.7.2
